training/pyproject.toml
[tool.poetry]
name = "training"
version = "0.1.0"
description = ""
authors = ["ICheered <tjeerd992@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
tensorflow = "^2.17.0"
scikit-learn = "^1.5.2"
pillow = "^10.4.0"
tensorflowjs = "^4.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
